requirements.txt
swibots
python-decouple
pyrogram
firebase-admin
aiohttp
requests-cache